refactor: UserEntity -> UserRequest
