+++ data/src/main/java/com/susu/data/model/request/UserRequest.kt
@@ -1,22 +1,16 @@
-package com.susu.data.model
+package com.susu.data.model.request
 
 import com.susu.core.model.User
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UserEntity(
+data class UserRequest(
     val name: String,
     val gender: String,
     val birth: Int,
 )
 
-fun UserEntity.toDomain() = User(
-    name = name,
-    gender = gender,
-    birth = birth,
-)
-
-fun User.toData() = UserEntity(
+fun User.toData() = UserRequest(
     name = name,
     gender = gender,
     birth = birth,
